requirements.txt
fastmcp>=2.0.0
google-genai>=1.0.0
Pillow>=10.0.0
PyYAML>=6.0
